+++ src/dp/swipe/target_iterator.h
@@ -261,7 +261,7 @@ struct AsyncTargetBuffer
 
 	int max_len() const {
 		int l = 0;
-		for (int i = 0; i < target_it.count; ++i)
+		for (size_t i = 0; i < target_it.count; ++i)
 			l = std::max(l, (int)target_it[i].seq.length());
 		return l;
 	}
